New structure for GPS module Device initialized in its own nodes, separate streams compartmentalized out
